Update Install  template.sh

+++ Bourne-Again Shell/Install  template.sh
@@ -1,44 +1,43 @@
 #!/bin/bash
-#Tutorial.....
+# CentOS 7.3 ...
 
 #定义...
-VARIABLE="VALUE"
+USERNAME="XXX"          #@27-31
 
 set -e
 set -x
 
-# source /etc/init.d/functions
-# echo_success,echo_failure,echo_passed,echo_warning
+source /etc/init.d/functions   # echo_success,echo_failure,echo_passed,echo_warning
 
 #身份检查
 if [ $(id -u) != "0" ]; then
-    echo "error: user must be an administrator"
+    echo "error: user must be an administrator"  echo_failure
     exit;
 fi
 
 #目录
 mkdir -p $..../etc
 
-#依赖
+#depend
 yum -y install epel-release gcc gcc-c++ cmake kernel-devel openssl openssl-devel
 
-#删除旧数据
+#erase old file and config ...
 rm -rf {配置目录,安装目录,解压目录,启动文件目录,其他目录...}
 
-#创建用户
-if ! id nginx &> /dev/null ; then
-    groupadd nginx
-    useradd -M -g nginx  nginx -s /sbin/nologin
+#add user
+if ! id ${USERNAME} &> /dev/null ; then
+    groupadd ${USERNAME}
+    useradd -M -g ${USERNAME} ${USERNAME:?'Undefined ...'} -s /sbin/nologin
 fi
 
-#判断是否有源码包
+#using local
 [ -s stunel-4.33.tar.gz ] || exit 1
 ......
 ......
-......
-#依内核数量并行
-NUM=$( awk '/processor/{NUM++};END{print NUM}' /proc/cpuinfo )
-if [ $NUM -gt 1 ] ;then
+
+#并行编译
+NUM=$( awk '/processor/{N++};END{print N}' /proc/cpuinfo )
+if [ $NUM -gt 1 ];then
     make -j $NUM
 else
     make
@@ -48,9 +47,17 @@ make install
 
 #关闭SELINUX与防火墙
 function disable_sec() {
-    setenforce 0 ; sed -i 's/^SELINUX=.*/SELINUX=disabled/g' /etc/sysconfig/selinux
-    systemctl disable firewalld #or firewall-cmd--permanent --add-port=XXX/tcp && firewall-cmd-reload
-    systemctl stop firewalld
+
+    if [ -x /usr/bin/systemctl ] ; then
+        #CentOS 7.X
+        setenforce 0 ; sed -i 's/^SELINUX=.*/SELINUX=disabled/g' /etc/sysconfig/selinux
+        systemctl disable firewalld #or firewall-cmd--permanent --add-port=XXX/tcp && firewall-cmd-reload
+        systemctl stop firewalld
+    else
+        #CentOS 6.X
+        chkconfig iptables off --level 235
+        service iptables stop
+        
 }
 
 disable_sec
@@ -58,6 +65,6 @@ disable_sec
 echo 'xxxxxxxxxxxxxx' >> /etc/rc.local
 chmod a+x /etc/rc.local
 
-echo -e "\nScript Execution Time： $SECONDS"
+echo -e "\033[32m\n\nScript Execution Time： ${SECONDS}s\033[0m"
 
 exit 0
